+++ src/nvim/statusline.c
@@ -1368,7 +1368,7 @@ int build_stl_str_hl(win_T *wp, char *out, size_t outlen, char *fmt, char *opt_n
 
     // An invalid item was specified.
     // Continue processing on the next character of the format string.
-    if (vim_strchr(STL_ALL, *fmt_p) == NULL) {
+    if (vim_strchr(STL_ALL, (uint8_t)(*fmt_p)) == NULL) {
       fmt_p++;
       continue;
     }
